cookbook/pocoflow-tool-search/requirements.txt
pocoflow>=0.2.0
python-dotenv>=1.0
click>=8.0
pyyaml>=6.0
duckduckgo-search>=7.5.2
